backend/requirements.txt
Django>=4
django-extensions
djangorestframework
channels
channels_redis
